+++ Engine/source/CameraManager.cpp
@@ -7,7 +7,7 @@
 void CameraManager::Init(glm::vec2 viewSize, CameraType type, float zoom, float angle)
 {
 	camera.SetCameraType(type);
-	camera.SetViewSize(viewSize.x, viewSize.y);
+	camera.SetViewSize(static_cast<int>(viewSize.x), static_cast<int>(viewSize.y));
 	camera.SetZoom(zoom);
 	camera.Rotate(angle);
